Extract native_transfers module to a file.
